requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==3.2.11
django-cors-headers==3.11.0
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
python-decouple==3.6